* Sepration of concern implemented to the venn digram controller functions
